# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT API
  description: |
    RESTful API for CozyVTT — a self-hosted, browser-based Virtual Tabletop platform.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
    Login via `/api/auth/login` to obtain a session cookie.

    ## MFA Flow
    If a user has MFA enabled, `/api/auth/login` returns `{ mfaRequired: true }` instead of a full session.
    The client must then call `/api/auth/mfa/verify-login` with a TOTP code or backup code to complete authentication.

    ## Authorization
    - **Platform roles**: `ADMIN` (full system access), `USER` (campaign access)
    - **Campaign roles**: `DM` (full campaign control), `PLAYER` (limited interaction), `SPECTATOR` (read-only)

    ## Rate Limiting
    - Global API limit: 300 requests per minute per IP (all `/api/*` routes)
    - Authentication endpoints: 5 requests per 15 minutes
    - MFA setup verification: 5 attempts per 15 minutes
    - MFA login verification: 5 attempts per 15 minutes (separate limit)
    - Asset uploads: 30 requests per minute per user (configurable via `ASSET_UPLOAD_RATE_LIMIT`)
    - Chat messages: 5 per minute (WebSocket)
    - Dice rolls: 30 per minute (WebSocket)
    - Token movement: 60 per second (WebSocket, throttled)

    Rate-limited responses return HTTP 429 with `RateLimit-*` standard headers.

    ## Real-time Events
...

  version: 1.0.0
  contact:
    name: CozyVTT
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/invitations:
    get:
      tags:
        - Invitations
      summary: List my pending invitations
      description: |
        Returns the authenticated user's pending campaign invitations (status
        `PENDING`), ordered newest first, with the inviting campaign's summary
        info included so the UI can render an invitation card without
        additional fetches.
      operationId: listMyInvitations
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Pending invitations.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    campaignId: { type: string, format: uuid }
                    userId: { type: string, format: uuid }
                    status:
                      type: string
                      enum: [PENDING, ACCEPTED, DECLINED, EXPIRED]
                    createdAt: { type: string, format: date-time }
                    expiresAt: { type: string, format: date-time, nullable: true }
                    campaign:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        description: { type: string, nullable: true }
                        gameSystem: { type: string, nullable: true }
                        status:
                          type: string
                          enum: [PREPARATION, ACTIVE, PAUSED, COMPLETED, ARCHIVED]
                        owner:
                          type: object
                          properties:
                            displayName: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/invitations/{id}/accept:
    post:
      tags:
        - Invitations
      summary: Accept a campaign invitation
      description: |
        Accept the invitation and join the campaign as a `PLAYER`. Optionally
        assigns one or more of the accepting user's characters to the campaign
        in the same transaction.

        Validation:
        - The invitation must belong to the authenticated user and be `PENDING`.
        - Expired invitations are rejected.
        - All `characterIds` must belong to the user and not already be assigned
          to a campaign.
        - If the campaign has a `gameSystem` set, each character's `gameSystem`
          must either be null or match the campaign's.

        On success, also broadcasts a `roster.updated` event over the WebSocket
        to all current campaign members.
      operationId: acceptInvitation
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/InvitationId'
      requestBody:
        required: false
        description: |
          Optional list of characters to bring into the campaign. Omit or send
          `{ "characterIds": [] }` to join without assigning any characters.
        content:
          application/json:
            schema:
              type: object
              properties:
                characterIds:
                  type: array
                  items: { type: string, format: uuid }
            example:
              characterIds:
                - 4fbd7b27-c5cd-4d6f-9c80-0f7f30c75c2f
      responses:
        '200':
          description: Invitation accepted; campaign membership created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Invitation accepted }
                  membership:
                    $ref: '#/components/schemas/CampaignMembership'
        '400':
          description: |
            Invitation already processed, expired, or character validation failed
            (not owned by user, already assigned, or game-system mismatch).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/invitations/{id}/decline:
    post:
      tags:
        - Invitations
      summary: Decline a campaign invitation
      description: |
        Marks the invitation as `DECLINED`. The invitation must belong to
        the authenticated user and currently be `PENDING`. Does not create
        any campaign membership.
      operationId: declineInvitation
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/InvitationId'
      responses:
        '200':
          description: Invitation declined.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Invitation declined }
        '400':
          description: Invitation already processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # ============================================
  # MESSAGES
  # ============================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
  parameters:
...
    InvitationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Invitation UUID
...
  schemas:
...
    CampaignMembership:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        role:
          type: string
          enum: [DM, PLAYER, SPECTATOR]
        characterIds:
          type: array
          items:
            type: string
            format: uuid
        joinedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Not authenticated or MFA pending.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication required.
    Forbidden:
      description: Forbidden - Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action.
    NotFound:
      description: Not Found - Resource does not exist or you cannot see it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Campaign not found.
...
    ServerError:
      description: Internal Server Error - Something went wrong on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InternalServerError
            message: An unexpected error occurred. Please try again later.